Added Math Evaluation Functionality
When an instructor views responses to 'free response' questions, it now
checks student responses against the instructor's 'answers'. It does
this even with unspecified variables (evaluating each variable at a
range of numerical values).

Also added a button on 'view_active_question_responses.php' to close a
question (so students can no longer respond to it).